Cosas sin sentido

+++ TP1/src/frsf/cidisi/exercise/tp1/search/actions/Avanzar.java
@@ -132,7 +132,7 @@ public class Avanzar extends SearchAction {
      */
     @Override
     public Double getCost() {
-        return new Double(0);
+        return new Double(1);
     }
 
     /**
