+++ TP1/src/frsf/cidisi/exercise/tp1/search/actions/GirarDer.java
@@ -76,7 +76,7 @@ public class GirarDer extends SearchAction {
      */
     @Override
     public Double getCost() {
-        return new Double(0);
+        return new Double(500);
     }
 
     /**
